requirements.txt
atlassian-python-api
simple-term-menu
pandas